Add other tx type tests
