+++ message-transact/src/tests/eip1559.rs
@@ -24,10 +24,11 @@ use crate::{mock::*, tests::*};
 use frame_support::pallet_prelude::Weight;
 use sp_core::U256;
 
-pub fn legacy_erc20_creation_unsigned_transaction() -> LegacyUnsignedTransaction {
-	LegacyUnsignedTransaction {
+pub fn eip1559_erc20_creation_unsigned_transaction() -> EIP1559UnsignedTransaction {
+	EIP1559UnsignedTransaction {
 		nonce: U256::zero(),
-		gas_price: U256::from(1),
+		max_priority_fee_per_gas: U256::from(1),
+		max_fee_per_gas: U256::from(1),
 		gas_limit: U256::from(0x100000),
 		action: ethereum::TransactionAction::Create,
 		value: U256::zero(),
@@ -43,8 +44,8 @@ fn test_dispatch_legacy_ethereum_transaction_works() {
 
 	ext.execute_with(|| {
 		let mock_message_id = [0; 4];
-		let unsigned_tx = legacy_erc20_creation_unsigned_transaction();
-		let t = unsigned_tx.sign(&alice.private_key);
+		let unsigned_tx = eip1559_erc20_creation_unsigned_transaction();
+		let t = unsigned_tx.sign(&alice.private_key, None);
 		let call = RuntimeCall::MessageTransact(crate::Call::message_transact { transaction: t });
 		let message = prepare_message(call);
 
@@ -77,10 +78,10 @@ fn test_dispatch_legacy_ethereum_transaction_weight_mismatch() {
 
 	ext.execute_with(|| {
 		let mock_message_id = [0; 4];
-		let mut unsigned_tx = legacy_erc20_creation_unsigned_transaction();
+		let mut unsigned_tx = eip1559_erc20_creation_unsigned_transaction();
 		// 62500001 * 16000 > 1_000_000_000_000
 		unsigned_tx.gas_limit = U256::from(62500001);
-		let t = unsigned_tx.sign(&alice.private_key);
+		let t = unsigned_tx.sign(&alice.private_key, None);
 		let call = RuntimeCall::MessageTransact(crate::Call::message_transact { transaction: t });
 		let message = prepare_message(call);
